move curl types to curl package and restructure curl package
