--- internal/curl/actions/set_headers.go
@@ -1,13 +1,0 @@
-package actions
-
-import (
-	"net/http"
-
-	"github.com/trco/wannabe/types"
-)
-
-func SetHeaders(request *http.Request, headers []types.Header) {
-	for _, v := range headers {
-		request.Header.Set(v.Key, v.Value)
-	}
-}
