+++ internal/curl/generate_curl_payload.go
@@ -1,25 +1,33 @@
-package actions
+package curl
 
 import (
 	"bytes"
 	"fmt"
 	"io"
 	"net/http"
-
-	"github.com/trco/wannabe/types"
 )
 
-func GenerateCurlPayload(request *http.Request) (types.CurlPayload, error) {
+type CurlPayload struct {
+	Scheme         string
+	HttpMethod     string
+	Host           string
+	Path           string
+	Query          map[string][]string
+	RequestHeaders map[string][]string
+	RequestBody    []byte
+}
+
+func GenerateCurlPayload(request *http.Request) (CurlPayload, error) {
 	body, err := io.ReadAll(request.Body)
 	if err != nil {
-		return types.CurlPayload{}, fmt.Errorf("PrepareGenerateCurlPayload: failed reading request body: %v", err)
+		return CurlPayload{}, fmt.Errorf("PrepareGenerateCurlPayload: failed reading request body: %v", err)
 	}
 	defer request.Body.Close()
 
 	// set body back to request
 	request.Body = io.NopCloser(bytes.NewBuffer(body))
 
-	curlPayload := types.CurlPayload{
+	curlPayload := CurlPayload{
 		Scheme:         request.URL.Scheme,
 		HttpMethod:     request.Method,
 		Host:           request.URL.Host,
