+++ internal/curl/prepare_curl.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 import (
 	"fmt"
