+++ internal/curl/prepare_request.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 import (
 	"bytes"
