+++ internal/curl/prepare_request_test.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 import (
 	"io"
