+++ internal/curl/prepare_url.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 func PrepareUrl(scheme string, host string, path string, query string) string {
 	if path == "/" {
