+++ internal/curl/prepare_url_test.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 import "testing"
