+++ internal/curl/process_http_method.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 func ProcessHttpMethod(httpMethod string) string {
 	return httpMethod
