+++ internal/curl/process_http_method_test.go
@@ -1,4 +1,4 @@
-package actions
+package curl
 
 import "testing"
