+++ internal/curl/process_path_test.go
@@ -1,15 +1,15 @@
-package actions
+package curl
 
 import (
 	"reflect"
 	"testing"
 
-	"github.com/trco/wannabe/types"
+	"github.com/trco/wannabe/internal/config"
 )
 
 type TestCaseProcessPath struct {
 	Path     string
-	Config   types.Path
+	Config   config.Path
 	Expected string
 }
 
@@ -19,15 +19,15 @@ func TestProcessPath(t *testing.T) {
 	tests := []struct {
 		name    string
 		path    string
-		config  types.Path
+		config  config.Path
 		want    string
 		wantErr string
 	}{
 		{
 			name: "with placeholder",
 			path: "/test1/test2/123456:test",
-			config: types.Path{
-				Wildcards: []types.WildcardIndex{{Index: &zero, Placeholder: "placeholder"}},
+			config: config.Path{
+				Wildcards: []config.WildcardIndex{{Index: &zero, Placeholder: "placeholder"}},
 			},
 			want:    "/placeholder/test2/123456:test",
 			wantErr: "",
@@ -35,8 +35,8 @@ func TestProcessPath(t *testing.T) {
 		{
 			name: "without placeholder",
 			path: "/test1/test2/123456:test",
-			config: types.Path{
-				Wildcards: []types.WildcardIndex{{Index: &zero}},
+			config: config.Path{
+				Wildcards: []config.WildcardIndex{{Index: &zero}},
 			},
 			want:    "/wannabe/test2/123456:test",
 			wantErr: "",
@@ -44,8 +44,8 @@ func TestProcessPath(t *testing.T) {
 		{
 			name: "with regex",
 			path: "/test1/test2/123456:test",
-			config: types.Path{
-				Regexes: []types.Regex{{Pattern: "(\\d+):test", Placeholder: "{id}:test"}},
+			config: config.Path{
+				Regexes: []config.Regex{{Pattern: "(\\d+):test", Placeholder: "{id}:test"}},
 			},
 			want:    "/test1/test2/{id}:test",
 			wantErr: "",
@@ -53,8 +53,8 @@ func TestProcessPath(t *testing.T) {
 		{
 			name: "empty string",
 			path: "",
-			config: types.Path{
-				Wildcards: []types.WildcardIndex{{Index: &zero}},
+			config: config.Path{
+				Wildcards: []config.WildcardIndex{{Index: &zero}},
 			},
 			want:    "",
 			wantErr: "",
@@ -62,8 +62,8 @@ func TestProcessPath(t *testing.T) {
 		{
 			name: "invalid regex",
 			path: "/test1/test2/123456:test",
-			config: types.Path{
-				Regexes: []types.Regex{{Pattern: "(?P<foo"}},
+			config: config.Path{
+				Regexes: []config.Regex{{Pattern: "(?P<foo"}},
 			},
 			want:    "",
 			wantErr: "ProcessPath: failed compiling regex: error parsing regexp: invalid named capture: `(?P<foo`",
