--- internal/types/curl.go
@@ -1,16 +1,0 @@
-package types
-
-type CurlPayload struct {
-	Scheme         string
-	HttpMethod     string
-	Host           string
-	Path           string
-	Query          map[string][]string
-	RequestHeaders map[string][]string
-	RequestBody    []byte
-}
-
-type Header struct {
-	Key   string
-	Value string
-}
